restruct proj
