+++ sentry/worker.py
@@ -5,35 +5,11 @@ from dataclasses import dataclass
 from datetime import timedelta
 
 import sentry_sdk
-from temporalio import activity, workflow
+from sentry.workflows import compose_greeting, GreetingWorkflow
 from temporalio.client import Client
 from temporalio.worker import Worker
 
 from sentry.interceptor import SentryInterceptor
-
-
-@dataclass
-class ComposeGreetingInput:
-    greeting: str
-    name: str
-
-
-@activity.defn
-async def compose_greeting(input: ComposeGreetingInput) -> str:
-    activity.logger.info("Running activity with parameter %s" % input)
-    return f"{input.greeting}, {input.name}!"
-
-
-@workflow.defn
-class GreetingWorkflow:
-    @workflow.run
-    async def run(self, name: str) -> str:
-        workflow.logger.info("Running workflow with parameter %s" % name)
-        return await workflow.execute_activity(
-            compose_greeting,
-            ComposeGreetingInput("Hello", name),
-            start_to_close_timeout=timedelta(seconds=10),
-        )
 
 
 async def main():
